intial create

+++ Web.APIs/Web.Domain/Entites/Property.cs
@@ -16,7 +16,7 @@ public class Property:BaseClass<int>
     public int Rooms { get; set; }
     public int Bathrooms { get; set; }
     public double Area { get; set; }
-    public decimal Price { get; set; }
+    public int Price { get; set; }
     public int Floor { get; set; }
     public PropertyType PropertyType { get; set; } //بيع او ايجار
     public string? Address { get; set; }
@@ -28,5 +28,6 @@ public class Property:BaseClass<int>
 
     public ICollection<PropertyImage> Images { get; set; } = new List<PropertyImage>();
     public ICollection<PropertyReview> Reviews { get; set; } = new List<PropertyReview>();
+
 
 }
